requirements.txt
pyspark>=3.3.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
databricks-connect>=13.0.0
great-expectations>=0.17.0
delta-spark>=2.4.0
pyyaml>=6.0.0
